spark-stream-ingest/src/main/python/requirements.txt
pandas==1.2.4
cortex-python==6.1.0
delta-spark==1.1.0
